no generic vector search core
