+++ src/frontend/src/optimizer/plan_node/logical_vector_search.rs
@@ -15,35 +15,129 @@
 use itertools::Itertools;
 use pretty_xmlish::{Pretty, XmlNode};
 use risingwave_common::bail;
+use risingwave_common::catalog::{Field, Schema};
+use risingwave_common::types::DataType;
 use risingwave_common::util::column_index_mapping::ColIndexMapping;
 use risingwave_common::util::sort_util::{ColumnOrder, OrderType};
 use risingwave_pb::common::PbDistanceType;
 
-use crate::PlanRef;
 use crate::expr::{
     ExprImpl, ExprRewriter, ExprType, ExprVisitor, FunctionCall, collect_input_refs,
 };
 use crate::optimizer::plan_node::expr_visitable::ExprVisitable;
-use crate::optimizer::plan_node::generic::{GenericPlanRef, TopNLimit, VectorSearch};
+use crate::optimizer::plan_node::generic::{GenericPlanNode, GenericPlanRef, TopNLimit};
 use crate::optimizer::plan_node::utils::{Distill, childless_record};
 use crate::optimizer::plan_node::{
     BatchProject, BatchTopN, ColPrunable, ColumnPruningContext, ExprRewritable, Logical,
     LogicalProject, PlanBase, PlanTreeNodeUnary, PredicatePushdown, PredicatePushdownContext,
     RewriteStreamContext, ToBatch, ToStream, ToStreamContext, gen_filter_and_pushdown, generic,
 };
-use crate::optimizer::property::Order;
-use crate::utils::Condition;
+use crate::optimizer::property::{FunctionalDependencySet, Order};
+use crate::utils::{ColIndexMappingRewriteExt, Condition};
+use crate::{OptimizerContextRef, PlanRef};
+
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+struct VectorSearchCore {
+    top_n: u64,
+    distance_type: PbDistanceType,
+    left: ExprImpl,
+    right: ExprImpl,
+    output_input_idx: Vec<usize>,
+    input: PlanRef,
+}
+
+impl VectorSearchCore {
+    pub(crate) fn clone_with_input(&self, input: PlanRef) -> Self {
+        Self {
+            top_n: self.top_n,
+            distance_type: self.distance_type,
+            left: self.left.clone(),
+            right: self.right.clone(),
+            output_input_idx: self.output_input_idx.clone(),
+            input,
+        }
+    }
+
+    pub(crate) fn rewrite_exprs(&mut self, r: &mut dyn ExprRewriter) {
+        self.left = r.rewrite_expr(self.left.clone());
+        self.right = r.rewrite_expr(self.right.clone());
+    }
+
+    pub(crate) fn visit_exprs(&self, v: &mut dyn ExprVisitor) {
+        v.visit_expr(&self.left);
+        v.visit_expr(&self.right);
+    }
+
+    pub(crate) fn i2o_mapping(&self) -> ColIndexMapping {
+        let mut mapping = vec![None; self.input.schema().len()];
+        for (output_idx, input_idx) in self.output_input_idx.iter().enumerate() {
+            mapping[*input_idx] = Some(output_idx);
+        }
+        ColIndexMapping::new(mapping, self.output_input_idx.len() + 1)
+    }
+}
+
+impl GenericPlanNode for VectorSearchCore {
+    fn functional_dependency(&self) -> FunctionalDependencySet {
+        self.i2o_mapping()
+            .rewrite_functional_dependency_set(self.input.functional_dependency().clone())
+    }
+
+    fn schema(&self) -> Schema {
+        let fields = self
+            .output_input_idx
+            .iter()
+            .map(|idx| self.input.schema()[*idx].clone())
+            .chain([Field::new("vector_distance", DataType::Float64)])
+            .collect();
+        Schema { fields }
+    }
+
+    fn stream_key(&self) -> Option<Vec<usize>> {
+        self.input.stream_key().and_then(|v| {
+            let i2o_mapping = self.i2o_mapping();
+            v.iter().map(|idx| i2o_mapping.try_map(*idx)).collect()
+        })
+    }
+
+    fn ctx(&self) -> OptimizerContextRef {
+        self.input.ctx()
+    }
+}
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub struct LogicalVectorSearch {
     pub base: PlanBase<Logical>,
-    core: VectorSearch<PlanRef>,
+    core: VectorSearchCore,
 }
 
 impl LogicalVectorSearch {
-    pub(crate) fn with_core(core: VectorSearch<PlanRef>) -> Self {
+    pub(crate) fn new(
+        top_n: u64,
+        distance_type: PbDistanceType,
+        left: ExprImpl,
+        right: ExprImpl,
+        output_input_idx: Vec<usize>,
+        input: PlanRef,
+    ) -> Self {
+        let core = VectorSearchCore {
+            top_n,
+            distance_type,
+            left,
+            right,
+            output_input_idx,
+            input,
+        };
+        Self::with_core(core)
+    }
+
+    fn with_core(core: VectorSearchCore) -> Self {
         let base = PlanBase::new_logical_with_core(&core);
         Self { base, core }
+    }
+
+    pub(crate) fn i2o_mapping(&self) -> ColIndexMapping {
+        self.core.i2o_mapping()
     }
 }
